feat: add validation to follow/unfollow and get recs api

+++ apps/recnet-api/src/modules/user/user.service.ts
@@ -118,9 +118,15 @@ export class UserService {
     userId: string,
     followingUserId: string
   ): Promise<void> {
-    // validate if the user exists
-    await this.userRepository.findUserById(userId);
-    await this.userRepository.findUserById(followingUserId);
+    // validate if the user exists and is activated
+    const followingUser =
+      await this.userRepository.findUserById(followingUserId);
+    if (!followingUser.isActivated) {
+      throw new RecnetError(
+        ErrorCode.ACCOUNT_NOT_ACTIVATED,
+        HttpStatus.BAD_REQUEST
+      );
+    }
 
     await this.followingRecordRepository.createFollowingRecord(
       userId,
@@ -132,9 +138,15 @@ export class UserService {
     userId: string,
     followingUserId: string
   ): Promise<void> {
-    // validate if the user exists
-    await this.userRepository.findUserById(userId);
-    await this.userRepository.findUserById(followingUserId);
+    // validate if the user exists and is activated
+    const followingUser =
+      await this.userRepository.findUserById(followingUserId);
+    if (!followingUser.isActivated) {
+      throw new RecnetError(
+        ErrorCode.ACCOUNT_NOT_ACTIVATED,
+        HttpStatus.BAD_REQUEST
+      );
+    }
 
     await this.followingRecordRepository.deleteFollowingRecord(
       userId,
